% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_next_event}
\alias{get_next_event}
\title{Get the next scheduled event}
\usage{
get_next_event()
}
\value{
Returns a dataframe with information about the next scheduled World Cup event.
}
\description{
The function returns the next scheduled event, filtered by WorldCup level and schedule status.
}
\examples{
get_next_event()
}
