% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_next_race}
\alias{get_next_race}
\title{Get the next scheduled race}
\usage{
get_next_race()
}
\value{
Returns a dataframe with information about the next scheduled World Cup race.
}
\description{
The function returns the next scheduled race, filtered by level and schedule status.
}
\examples{
get_next_event()
}
