% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_next_race_plus_n}
\alias{get_next_race_plus_n}
\title{Get the after n next scheduled race}
\usage{
get_next_race_plus_n(n = 0)
}
\arguments{
\item{n}{Jump to n after next race. Default is 0.}
}
\value{
Returns a dataframe with information about the next scheduled World Cup race. Note: only for the
event.
}
\description{
The function returns the next scheduled race plus n, filtered by level and schedule status.
For the moment, team events are omitted here.
}
\examples{
get_next_race_plus_n()
}
