% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_previous_seasons}
\alias{get_previous_seasons}
\title{Returns the previous seasons of a given date.}
\usage{
get_previous_seasons(date, minus_years = 1)
}
\arguments{
\item{date}{A date in character, numeric, or date format.}

\item{minus_years}{Number of years to go back from the input date (default is 1)}
}
\value{
A character string representing the season in the format of 'YYYY',
where the first two digits represent the starting year and the last two digits
represent the ending year of the season.
}
\description{
Returns the previous seasons of a given date.
}
\examples{
get_previous_seasons("2021-06-01", 3)
}
