% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_race_prediction.R
\name{get_race_prediction}
\alias{get_race_prediction}
\title{Get predictions of a biathlon race}
\usage{
get_race_prediction(
  race_id,
  gbm_model,
  n_long = 26,
  n_long_disc = 10,
  n_short = 3,
  n_slope = 3
)
}
\arguments{
\item{race_id}{A string representing the race id.}

\item{gbm_model}{A gbm model object.}

\item{n_long}{The number races to consider for computing long-term features. Default is 26.}

\item{n_long_disc}{The number races to consider for computing long-term, discipline-specific features. Default is 10.}

\item{n_short}{The number races to consider for computing short-term features. Default is 3.}

\item{n_slope}{The number races to consider for computing the slope of the trend in previous performances.}
}
\value{
Returns a dataframe with predictions of the race.
}
\description{
The function returns predictions of a Biathlon race, by calling the update_features function and a gbm model.
}
\examples{
data("gbm.ndcg5")
get_race_prediction("BT2122SWRLCP04SWSP", gbm.ndcg5)
}
