% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_season}
\alias{get_season}
\title{Returns the season of a given date.}
\usage{
get_season(date)
}
\arguments{
\item{date}{A date in character, numeric, or date format.}
}
\value{
A character string representing the season in the format of 'YYYY',
where the first two digits represent the starting year and the last two digits
represent the ending year of the season.
}
\description{
Returns the season of a given date.
}
\examples{
get_season("2021-06-01")
}
