% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_features.R
\name{update_features}
\alias{update_features}
\title{Update features for a list of athletes}
\usage{
update_features(
  athlete_list,
  cutoff_datetime = NULL,
  n_long = 26,
  n_long_disc = 10,
  n_short = 3,
  n_slope = 3
)
}
\arguments{
\item{athlete_list}{A list of athlete IBUIds}

\item{cutoff_datetime}{A date time up to which the features need to be updated. Default is current date time}

\item{n_long}{number of long races to consider in the features}

\item{n_long_disc}{number of races of each discipline to consider in the features}

\item{n_short}{number of short races to consider in the features}

\item{n_slope}{number of slope races to consider in the features}
}
\value{
A dataframe with the updated features for each athlete
}
\description{
Update features for a list of athletes by getting all the
results of the athletes from the biathlonResults API wrapper. ... ADD DESCRIPTION
}
\examples{
update_features(c("BTNOR11605199301"), n_long = 2, n_slope = 2)
}
